+++ client/src/components/Pagination.tsx
@@ -17,20 +17,21 @@
  */
 
 import cx from "classnames";
-import { memo } from "react";
+import { memo, useCallback } from "react";
 import {
   ChevronDoubleLeft,
   ChevronDoubleRight,
   ChevronLeft,
   ChevronRight,
 } from "react-bootstrap-icons";
+import { Link, useLocation, type To } from "react-router";
 import { PaginationItem, PaginationLink } from "reactstrap";
 
 interface PaginationProps {
   ariaLabel?: string;
   className?: string;
   currentPage: number;
-  onPageChange: (pageNumber: number) => void;
+  pageQueryParam: string;
   perPage: number;
   showDescription?: boolean;
   totalInPage?: number;
@@ -41,7 +42,7 @@ export default function Pagination({
   ariaLabel,
   className: className_,
   currentPage,
-  onPageChange,
+  pageQueryParam,
   perPage,
   showDescription,
   totalInPage,
@@ -61,7 +62,7 @@ export default function Pagination({
         ariaLabel={ariaLabel}
         innerClassName={className}
         itemsCountPerPage={perPage}
-        onChange={onPageChange}
+        pageQueryParam={pageQueryParam}
         totalItemsCount={totalItems}
       />
       {showDescription && totalInPage && (
@@ -82,7 +83,7 @@ interface PaginationNavProps {
   innerClassName?: string;
   itemsCountPerPage: number;
   maxPages?: number;
-  onChange: (pageNumber: number) => void;
+  pageQueryParam: string;
   totalItemsCount: number;
 }
 
@@ -91,10 +92,26 @@ const PaginationNav = memo(function PaginationNav({
   ariaLabel = "page navigation",
   innerClassName,
   itemsCountPerPage,
-  onChange,
+  pageQueryParam,
   maxPages = 5,
   totalItemsCount,
 }: PaginationNavProps) {
+  const { hash, pathname, search } = useLocation();
+
+  const getPageLink = useCallback(
+    (page: number) => {
+      const newSearch = new URLSearchParams(search);
+      newSearch.set(pageQueryParam, `${page}`);
+      const to: To = {
+        hash: hash,
+        pathname: pathname,
+        search: newSearch.toString(),
+      };
+      return to;
+    },
+    [hash, pageQueryParam, pathname, search]
+  );
+
   if (itemsCountPerPage === 0 || totalItemsCount === 0 || maxPages <= 0) {
     return null;
   }
@@ -117,7 +134,7 @@ const PaginationNav = memo(function PaginationNav({
       <PaginationElement
         linkClassName="px-2"
         ariaLabel="first page"
-        onClick={() => onChange(1)}
+        to={getPageLink(1)}
         key="first"
       >
         <ChevronDoubleLeft className="bi" />
@@ -131,7 +148,7 @@ const PaginationNav = memo(function PaginationNav({
       <PaginationElement
         linkClassName="px-2"
         ariaLabel="previous page"
-        onClick={() => onChange(activePage - 1)}
+        to={getPageLink(activePage - 1)}
         key="prev"
       >
         <ChevronLeft className="bi" />
@@ -144,7 +161,7 @@ const PaginationNav = memo(function PaginationNav({
     pages.push(
       <PaginationElement
         ariaLabel={`page ${pageNumber} of ${totalPages}`}
-        onClick={() => onChange(pageNumber)}
+        to={getPageLink(pageNumber)}
         className={cx(pageNumber === activePage && "active")}
         key={`page-${pageNumber}`}
       >
@@ -159,7 +176,7 @@ const PaginationNav = memo(function PaginationNav({
       <PaginationElement
         linkClassName="px-2"
         ariaLabel="next page"
-        onClick={() => onChange(activePage + 1)}
+        to={getPageLink(activePage + 1)}
         key="next"
       >
         <ChevronRight className="bi" />
@@ -173,7 +190,7 @@ const PaginationNav = memo(function PaginationNav({
       <PaginationElement
         linkClassName="px-2"
         ariaLabel="last page"
-        onClick={() => onChange(totalPages)}
+        to={getPageLink(totalPages)}
         key="last"
       >
         <ChevronDoubleRight className="bi" />
@@ -193,26 +210,22 @@ interface PaginationElementProps {
   children?: React.ReactNode;
   className?: string;
   linkClassName?: string;
-  onClick: () => void;
+  to: To;
 }
 function PaginationElement({
   ariaLabel,
-  children,
-  className,
   linkClassName,
-  onClick,
+  className,
+  children,
+  to,
 }: PaginationElementProps) {
   return (
     <PaginationItem className={className}>
       <PaginationLink
+        tag={Link}
+        to={to}
         aria-label={ariaLabel}
         className={linkClassName}
-        href="#"
-        // This should be converted to use Link and allow for smooth navigation
-        onClick={(e) => {
-          e.preventDefault();
-          onClick();
-        }}
       >
         {children}
       </PaginationLink>
